requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
langgraph==0.6.5
pandas==2.3.2
duckdb==1.4.0
pydantic==2.11.7
pydantic-settings==2.10.1
streamlit==1.49.1
requests==2.32.5
google-genai==1.30.0
openai==1.107.1
pytest==8.4.1
httpx==0.28.1
